+++ src/diag_cluster2d.h
@@ -33,7 +33,7 @@ class DiagCluster2d : public Diag {
   void dump_clusters(double);
   void dump_clusters_detailed(double);
   void generate_clusters();
-  void add_cluster(int, int, int, int*);
+  void add_cluster(int, double, int, double*);
   void free_clustlist();
 
   int** cluster_ids;
